# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FinTech Personal Finance Dashboard API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface SignupRequest {
  name: string;
  email: string;
  password: string;
  salary: number;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export interface User {
  id: number;
  name: string;
  email: string;
  salary: number;
}

export interface AuthResponse {
  user: User;
  message: string;
}

export interface Expense {
  id: number;
  userId: number;
  description: string;
  category: string;
  amount: number;
  date: string;
}

export type CreateExpenseRequestCategory =
  (typeof CreateExpenseRequestCategory)[keyof typeof CreateExpenseRequestCategory];

export const CreateExpenseRequestCategory = {
  Food: "Food",
  Transport: "Transport",
  Shopping: "Shopping",
  Bills: "Bills",
  Entertainment: "Entertainment",
  Other: "Other",
} as const;

export interface CreateExpenseRequest {
  description: string;
  category: CreateExpenseRequestCategory;
  amount: number;
  date: string;
}

export interface SavingsGoal {
  id: number;
  userId: number;
  goalName: string;
  targetAmount: number;
  months: number;
  monthlySaving: number;
}

export interface CreateSavingsGoalRequest {
  goalName: string;
  targetAmount: number;
  months: number;
}

export interface UpdateSalaryRequest {
  salary: number;
}

export interface DashboardSummary {
  salary: number;
  totalExpenses: number;
  netSavings: number;
  savingsRate: number;
  recentExpenses: Expense[];
}