# lib/api-zod/src/generated/types/createExpenseRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FinTech Personal Finance Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateExpenseRequestCategory } from "./createExpenseRequestCategory";

export interface CreateExpenseRequest {
  description: string;
  category: CreateExpenseRequestCategory;
  amount: number;
  date: string;
}