# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FinTech Personal Finance Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { Expense } from "./expense";

export interface DashboardSummary {
  salary: number;
  totalExpenses: number;
  netSavings: number;
  savingsRate: number;
  recentExpenses: Expense[];
}