# lib/api-zod/src/generated/types/savingsGoal.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FinTech Personal Finance Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface SavingsGoal {
  id: number;
  userId: number;
  goalName: string;
  targetAmount: number;
  months: number;
  monthlySaving: number;
}