# lib/api-zod/src/generated/types/signupRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FinTech Personal Finance Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface SignupRequest {
  name: string;
  email: string;
  password: string;
  salary: number;
}